Huggingface Token UI (#285)
- [x] Pop-up a huggingface UI that allows user to enter token for models
that are gated on HF.
- [x] Integrate in `server.py` and catch the error when a model cannot
be downloaded / is gated.
- [x] Support token verification
```
You are trying to access a gated repo.
Make sure to have access to it at https://huggingface.co/meta-llama/Llama-3.2-1B.
401 Client Error. (Request ID: Root=1-674673a3-0848e79727a7a1820ce034f3;40781733-93e2-4a7d-8bb4-dabf15b0f295)

Cannot access gated repo for url https://huggingface.co/meta-llama/Llama-3.2-1B/resolve/main/config.json.
Access to model meta-llama/Llama-3.2-1B is restricted. You must have access to it and be authenticated to access it. Please log in.
```


![image](https://github.com/user-attachments/assets/7d1797da-7eb4-41a2-8b0e-0c3d2874b6d8)


![image](https://github.com/user-attachments/assets/80ef3e7d-3187-4696-849f-a7da1781475f)