Release.py for GAIA (#371)
A version of release.py borrowed from lemonade project and modified to
support the release process for the GAIA OSS project.
- [x] Added and modified release.py, run: `python release.py
..\gaia_internal` (see
[release.md](https://github.com/aigdat/gaia/blob/7bdff2ae6de34106c012768a3a56ac00f5af1489/nda/docs/release.md)
for details)
- [x] Created a `docs.py` utility that updates all references of
`aigdata/gaia` into `amd/gaia`.
- [x] Integrate github link updates across documentation in `release.py`
- [x] Moved `nimbys.md` documentation to `nda/docs` folder
- [x] Created a
[release.md](https://github.com/aigdat/gaia/blob/7bdff2ae6de34106c012768a3a56ac00f5af1489/nda/docs/release.md)
document to outline the OSS release process

It would be great if someone else tries to go through the release
process next round to test out the tools and documentation.

cc: @vgodsoe, @itomek

---------

Co-authored-by: Victoria Godsoe <victoria.godsoe@amd.com>